osclib/list_command.py: Show details also for requests targeted to packages not in rings.
This commit implements the ability to show for packages that do not
belong to any ring the same details that are shown for packages that do
live in rings.  Due to concerns about compatibility and volume of the
output the feature is hidden behind a new --adi-details command
line flag.